+++ NINA.Core.WPF/ViewModel/Equipment/WeatherData/WeatherDataVM.cs
@@ -315,6 +315,9 @@ namespace NINA.WPF.Base.ViewModel.Equipment.WeatherData {
                 RaisePropertyChanged();
             }
         }
+        public IDevice GetDevice() {
+            return WeatherData;
+        }
         public IDeviceChooserVM DeviceChooserVM { get; set; }
 
         private DeviceUpdateTimer updateTimer;